small refactor of ccta modules
