+++ ROADMAP.MD
@@ -23,8 +23,8 @@
 - [x] use the cleanup_section_points() also on LCA points
 - [x] Shrink/expand coronary to match intravascular images
 - [x] Shrink/Expand CCTA to best match aortic wall
-- [] Write a more stable clean up function by going along every face of a vertix adn then check if most vertices are from the same type
-- [] Remove points inside of anomalous region, connect remaining points to proximal and distal points, without resampling of the anomalous region
+- [x] Write a more stable clean up function by going along every face of a vertix adn then check if most vertices are from the same type
+- [x] Remove points inside of anomalous region, connect remaining points to proximal and distal points, without resampling of the anomalous region
 - [] Create ostium frames for smoother connection to aorta (define range cl points, and function of lumen increase)
 - [] Write Pytests for the new functions
 - [] Write a more stable function to find intramural points on aorta (for now works with excluded RCA points)
